requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
celery
playwright
lxml
beautifulsoup4
python-dotenv
PyMuPDF
python-docx
textract
faiss-cpu
scikit-learn
tesseract
pytesseract